+++ idl/src/codegen/typescript/mod.rs
@@ -1,0 +1,6 @@
+mod shared;
+
+pub use shared::{
+    client_dependency_version, generate_package_json, generate_ts_client, generate_ts_client_kit,
+    TsTarget,
+};
